requirements.txt
bcrypt==3.2.0
cffi==1.15.0
click==8.1.2
colorama==0.4.4
dnspython==2.2.1
email-validator==1.1.3
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
flask-kafka==0.0.3
flask-marshmallow==0.14.0
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
kafka-python==2.0.2
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
mongoengine==0.24.1
packaging==21.3
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pymongo==4.1.0
PyMySQL==1.0.2
pyparsing==3.0.7
python-decouple==3.6
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.34
Werkzeug==2.1.1
WTForms==3.0.1
